church_finance_backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
pandas==1.3.3
openpyxl==3.0.8
python-dotenv==0.19.0
alembic==1.7.5
pyjwt==2.1.0
email-validator==1.1.3
msal==1.20.0
requests==2.28.1